refactor: move validation into utils
